requirements-airflow.txt
apache-airflow==2.9.3
pyspark==3.5.1
